# swagger.yaml
openapi: 3.0.3
info:
  title: Hotel Room Booking System API
  description: |
    A comprehensive REST API for managing hotel room bookings, including booking management,
    messaging with hotel staff, room information, and authentication.
    
    **User Roles:**
    - **Admin**: Full access to all endpoints, can manage bookings, rooms, and read messages
    - **User**: Can create bookings, view rooms, check availability, and send messages
    - **Guest**: Can view rooms and make bookings without authentication
    
    **Features:**
    - Secure JWT-based authentication
    - Room availability management
    - Booking lifecycle management
    - Payment integration support
    - Guest communication system
    - Cancellation policy management
  version: 1.0.0
  contact:
    name: API Support
    email: support@hotelbooking.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.hotelbooking.com/v1
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /payments:
    post:
      tags:
        - Payment
      summary: Create a payment
      description: Create a payment for a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentCreateRequest'
      responses:
        '201':
          description: Payment initiated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          description: Invalid payment request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /payments/{id}:
    get:
      tags:
        - Payment
      summary: Get payment details
      description: Retrieve payment information
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Payment ID
      responses:
        '200':
          description: Payment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '404':
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /payments/booking/{bookingId}:
    get:
      tags:
        - Payment
      summary: Get payments for a booking
      description: Retrieve all payments associated with a booking
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Booking ID
      responses:
        '200':
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaymentResponse'
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /payments/{id}/refund:
    post:
      tags:
        - Payment
      summary: Refund a payment (Admin only)
      description: Process a refund for a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Payment ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
                - reason
              properties:
                amount:
                  type: number
                  format: double
                  example: 150.00
                reason:
                  type: string
                  maxLength: 500
                  example: Booking cancelled within refund period
      responses:
        '200':
          description: Refund processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          description: Invalid refund request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Guest API
...
  /rooms:
    post:
      tags:
        - Room
      summary: Create a new room (Admin only)
      description: Create a new room in the system - admin access required
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomCreateRequest'
      responses:
        '201':
          description: Room created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Room number already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Room
      summary: Get all rooms
      description: Retrieve list of all rooms (accessible by both users and admins)
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
          description: Page number
        - name: size
          in: query
          schema:
            type: integer
            default: 10
          description: Page size
        - name: type
          in: query
          schema:
            type: string
            enum: [SINGLE, DOUBLE, SUITE, DELUXE]
          description: Filter by room type
        - name: minPrice
          in: query
          schema:
            type: number
            format: double
          description: Minimum price per night
        - name: maxPrice
          in: query
          schema:
            type: number
            format: double
          description: Maximum price per night
      responses:
        '200':
          description: List of rooms
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomPageResponse'
  /rooms/{id}:
    get:
      tags:
        - Room
      summary: Get room details
      description: Retrieve detailed information about a specific room
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Room ID
      responses:
        '200':
          description: Room details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
        '404':
          description: Room not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Room
      summary: Update room (Admin only)
      description: Update room details - admin access required
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Room ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomUpdateRequest'
      responses:
        '200':
          description: Room updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Room not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Room
      summary: Delete room (Admin only)
      description: Delete a room - admin access required
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Room ID
      responses:
        '204':
          description: Room deleted successfully
        '403':
          description: Forbidden - admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Room not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Cannot delete room with existing bookings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RoomCreateRequest:
      type: object
      required:
        - roomNumber
        - type
        - pricePerNight
        - capacity
      properties:
        roomNumber:
          type: string
          example: "101"
        type:
          type: string
          enum: [SINGLE, DOUBLE, SUITE, DELUXE]
          example: DOUBLE
        pricePerNight:
          type: number
          format: double
          minimum: 0
          example: 150.00
        capacity:
          type: integer
          minimum: 1
          maximum: 10
          example: 2
        description:
          type: string
          maxLength: 1000
          example: Spacious double room with city view
        amenities:
          type: array
          items:
            type: string
          example: ["WiFi", "TV", "Air Conditioning", "Mini Bar"]
        imageUrl:
          type: string
          format: uri
          example: https://example.com/images/room-101.jpg
    RoomUpdateRequest:
      type: object
      properties:
        type:
          type: string
          enum: [SINGLE, DOUBLE, SUITE, DELUXE]
          example: DOUBLE
        pricePerNight:
          type: number
          format: double
          minimum: 0
          example: 150.00
        capacity:
          type: integer
          minimum: 1
          maximum: 10
          example: 2
        description:
          type: string
          maxLength: 1000
          example: Spacious double room with city view
        amenities:
          type: array
          items:
            type: string
          example: ["WiFi", "TV", "Air Conditioning", "Mini Bar"]
        imageUrl:
          type: string
          format: uri
          example: https://example.com/images/room-101.jpg
        isAvailable:
          type: boolean
          example: true
    RoomResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        roomNumber:
          type: string
          example: "101"
        type:
          type: string
          enum: [SINGLE, DOUBLE, SUITE, DELUXE]
          example: DOUBLE
        pricePerNight:
          type: number
          format: double
          example: 150.00
        capacity:
          type: integer
          example: 2
        size:
          type: number
          format: double
          example: 35.5
          description: Room size in square meters
        bedType:
          type: string
          example: King Size
        description:
          type: string
          example: Spacious double room with city view
        amenities:
          type: array
          items:
            type: string
          example: ["WiFi", "TV", "Air Conditioning", "Mini Bar", "Safe", "Coffee Maker"]
        images:
          type: array
          items:
            type: string
            format: uri
          example: 
            - https://example.com/images/room-101-1.jpg
            - https://example.com/images/room-101-2.jpg
        imageUrl:
          type: string
          example: https://example.com/images/room-101.jpg
          description: Primary image (deprecated, use images array)
        floor:
          type: integer
          example: 1
          description: Floor number where room is located
        view:
          type: string
          example: City View
        smokingAllowed:
          type: boolean
          example: false
        petFriendly:
          type: boolean
          example: false
        isAvailable:
          type: boolean
          example: true
        maintenanceStatus:
          type: string
          enum: [AVAILABLE, UNDER_MAINTENANCE, OUT_OF_ORDER]
          example: AVAILABLE
        rating:
          type: number
          format: double
          example: 4.5
          description: Average guest rating
        reviewCount:
          type: integer
          example: 127
        createdAt:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-10-07T10:30:00Z"
    RoomPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/RoomResponse'
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 10
        totalElements:
          type: integer
          format: int64
          example: 50
        totalPages:
          type: integer
          example: 5

    # Message Schemas
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2025-10-07T10:30:00Z"
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Invalid input data
        path:
          type: string
          example: /api/v1/bookings
        details:
          type: array
          items:
            type: string
          example: ["checkInDate must be a future date", "guestEmail must be a valid email"]

    # Payment Schemas
    PaymentCreateRequest:
      type: object
      required:
        - bookingId
        - amount
        - paymentMethod
      properties:
        bookingId:
          type: integer
          format: int64
          example: 1
        amount:
          type: number
          format: double
          minimum: 0.01
          example: 150.00
        paymentMethod:
          type: string
          enum: [CREDIT_CARD, DEBIT_CARD, PAYPAL, BANK_TRANSFER, CASH]
          example: CREDIT_CARD
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          example: USD
          default: USD
        cardDetails:
          type: object
          properties:
            cardNumber:
              type: string
              pattern: '^\d{16}$'
              example: "4532015112830366"
            cardHolderName:
              type: string
              example: John Doe
            expiryMonth:
              type: integer
              minimum: 1
              maximum: 12
              example: 12
            expiryYear:
              type: integer
              example: 2027
            cvv:
              type: string
              pattern: '^\d{3,4}$'
              example: "123"
        billingAddress:
          $ref: '#/components/schemas/Address'
    PaymentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        bookingId:
          type: integer
          format: int64
          example: 1
        transactionId:
          type: string
          example: TXN-2025-00001
        amount:
          type: number
          format: double
          example: 150.00
        currency:
          type: string
          example: USD
        paymentMethod:
          type: string
          enum: [CREDIT_CARD, DEBIT_CARD, PAYPAL, BANK_TRANSFER, CASH]
          example: CREDIT_CARD
        status:
          type: string
          enum: [PENDING, PROCESSING, COMPLETED, FAILED, REFUNDED, CANCELLED]
          example: COMPLETED
        paymentGatewayResponse:
          type: string
          example: Transaction approved
        refundAmount:
          type: number
          format: double
          example: 0.00
        createdAt:
          type: string
          format: date-time
          example: "2025-10-07T10:30:00Z"
        completedAt:
          type: string
          format: date-time
          example: "2025-10-07T10:31:00Z"

    # Guest Schemas
...
    Address:
      type: object
      properties:
        street:
          type: string
          example: 123 Main Street
        city:
          type: string
          example: New York
        state:
          type: string
          example: NY
        postalCode:
          type: string
          example: "10001"
        country:
          type: string
          example: USA
...